[Refac] 목데이터 제거
